Разделяет код из main.js на отдельные модули
